+++ package.json
@@ -25,6 +25,7 @@
   "dependencies": {
     "@node-minify/core": "^6.1.0",
     "@node-minify/uglify-es": "^6.1.0",
+    "browserify": "^17.0.0",
     "ejs": "^3.1.5",
     "express": "^4.17.1",
     "node-sri": "^1.1.1",
