+++ pages/game.html
@@ -36,6 +36,7 @@
             <li>Touch hold: Charge jump</li>
             <li>Drag: Squid moves to finger</li>
         </ul>
+        <p>Find a bug or have an idea? <a href="https://github.com/shatindle/rsplatoon-minigame/issues" target="_blank">Add an issue!</a></p>
         
         <script src="/js/game.min.js" integrity="##jsintegrity##"></script>
     </body>
