+++ src/animations.js
@@ -1,0 +1,19 @@
+const splashAnimation = {
+    x: 0,
+    y: 100,
+    width: 120,
+    height: 57,
+    started: null
+};
+
+const arrowAnimation = {
+    x: 0,
+    y: 0,
+    width: 50,
+    height: 38
+};
+
+module.exports = {
+    splash: splashAnimation,
+    arrow: arrowAnimation
+};
