+++ src/init.js
@@ -1,0 +1,8 @@
+module.exports = () => {
+    var requestAnimationFrame = window.requestAnimationFrame ||
+        window.mozRequestAnimationFrame ||
+        window.webkitRequestAnimationFrame ||
+        window.msRequestAnimationFrame;
+
+    window.requestAnimationFrame = requestAnimationFrame;
+};
