+++ src/isTouchDevice.js
@@ -1,0 +1,3 @@
+const isTouch = 'ontouchstart' in document.documentElement;
+
+module.exports = isTouch;
