+++ src/main.js
@@ -1,136 +1,3 @@
-(function(){function r(e,n,t){function o(i,f){if(!n[i]){if(!e[i]){var c="function"==typeof require&&require;if(!f&&c)return c(i,!0);if(u)return u(i,!0);var a=new Error("Cannot find module '"+i+"'");throw a.code="MODULE_NOT_FOUND",a}var p=n[i]={exports:{}};e[i][0].call(p.exports,function(r){var n=e[i][1][r];return o(n||r)},p,p.exports,r,e,n,t)}return n[i].exports}for(var u="function"==typeof require&&require,i=0;i<t.length;i++)o(t[i]);return o}return r})()({1:[function(require,module,exports){
-const splashAnimation = {
-    x: 0,
-    y: 100,
-    width: 120,
-    height: 57,
-    started: null
-};
-
-const arrowAnimation = {
-    x: 0,
-    y: 0,
-    width: 50,
-    height: 38
-};
-
-module.exports = {
-    splash: splashAnimation,
-    arrow: arrowAnimation
-};
-},{}],2:[function(require,module,exports){
-const isTouch = require("./isTouchDevice");
-
-const width = 300,
-    height = isTouch ? 450 : 500,
-    canvasHeight = 500,
-    midHeight = height / 2,
-    friction = 0.8,
-    gravity = isTouch ? 0.13 : 0.2;
-
-module.exports = {
-    width: width,
-    height: height,
-    canvasHeight: canvasHeight,
-    midHeight: midHeight,
-    friction: friction,
-    gravity: gravity
-};
-},{"./isTouchDevice":6}],3:[function(require,module,exports){
-var lastFrame = Date.now() - 1;
-var thisFrame = Date.now();
-
-function show(ctx, debug) {
-    var delta = Math.floor(1000 / (thisFrame - lastFrame));
-
-    lastFrame = thisFrame;
-
-    if (debug) {
-        ctx.font = '12px SplatRegular';
-        ctx.fillStyle = "black";
-        ctx.fillText(delta + " fps", 10, 16);
-    }
-}
-
-function update() {
-    thisFrame = Date.now();
-}
-
-function currentFrame() {
-    return thisFrame;
-}
-
-module.exports = {
-    update: update,
-    currentFrame: currentFrame,
-    show: show
-};
-},{}],4:[function(require,module,exports){
-const hexGradients = [
-    "#F0ECDB", // 50
-    "#F0F0DB",
-    "#ECF0DB",
-    "#E9F0DB",
-    "#E6F0DB",
-    "#E2F0DB",
-    "#DFF0DB",
-    "#DBF0DB",
-    "#DBF0DF",
-    "#DBF0E2",
-    "#DBF0E6",
-    "#DBF0E9",
-    "#DBF0EC",
-    "#DBF0F0",
-    "#DBECF0",
-    "#DBE9F0",
-    "#DBE6F0",
-    "#DBE2F0",
-    "#DBDFF0",
-    "#DBDBF0",
-    "#DFDBF0",
-    "#E2DBF0",
-    "#E6DBF0",
-    "#E9DBF0",
-    "#ECDBF0",
-    "#F0DBF0",
-    "#F0DBEC",
-    "#F0DBE9",
-    "#F0DBE6",
-    "#F0DBE2",
-    "#F0DBDF",
-    "#F0DBDB",
-    "#F0DFDB",
-    "#F0E2DB",
-    "#F0E6DB",
-    "#F0E9DB"
-];
-
-function getGradient(score) {
-    // 9999 / gradients = distance per gradient
-    // floor(score / distance) = item (with overflow)
-    // item % gradients = item to use
-    var dG = 300  / hexGradients.length;
-    var iG = Math.floor(score / dG);
-    var iI = iG % hexGradients.length;
-
-    return hexGradients[iI];
-}
-
-module.exports = getGradient;
-},{}],5:[function(require,module,exports){
-module.exports = () => {
-    var requestAnimationFrame = window.requestAnimationFrame ||
-        window.mozRequestAnimationFrame ||
-        window.webkitRequestAnimationFrame ||
-        window.msRequestAnimationFrame;
-
-    window.requestAnimationFrame = requestAnimationFrame;
-};
-},{}],6:[function(require,module,exports){
-const isTouch = 'ontouchstart' in document.documentElement;
-
-module.exports = isTouch;
-},{}],7:[function(require,module,exports){
 "use strict";
 
 var debug = false;
@@ -658,230 +525,3 @@ if (isTouch) {
 }
 
 window.share = share;
-},{"./animations":1,"./dimensions":2,"./frameData":3,"./hexGradients":4,"./init":5,"./isTouchDevice":6,"./page":8,"./player":9,"./screenObjects":10,"./sprites":11,"./userInput":12}],8:[function(require,module,exports){
-const dimensions = require("./dimensions");
-
-var canvas = document.getElementById("canvas"),
-    ctx = canvas.getContext("2d"),
-    bottomIsDeath = false;
-
-canvas.width = dimensions.width;
-canvas.height = dimensions.canvasHeight;
-
-module.exports = {
-    canvas: canvas,
-    ctx: ctx,
-    bottomIsDeath: bottomIsDeath
-};
-},{"./dimensions":2}],9:[function(require,module,exports){
-const dimensions = require("./dimensions");
-
-const player = {
-    x: dimensions.width / 2,
-    y: dimensions.height - 5,
-    width: 30,
-    height: 35,
-    speed: 4,
-    velX: 0,
-    velY: 0,
-    jumping: false,
-    jumpHold: null,
-    jumpHoldTime: null,
-    ghost: 0, // 0 = no ghost, 1 = left ghost
-    grounded: false,
-    dead: false,
-    score: 0
-};
-
-module.exports = player;
-},{"./dimensions":2}],10:[function(require,module,exports){
-const dimensions = require("./dimensions");
-
-const boxTypes = {
-    ROCK: 0,
-    SUCTION: 1,
-    ARROW: 2
-};
-
-const boxTypeOptions = [
-    boxTypes.SUCTION,
-    boxTypes.SUCTION,
-    boxTypes.SUCTION,
-    boxTypes.SUCTION,
-    boxTypes.ARROW
-];
-
-const boxes = [];
-
-function generatePlatform(init) {
-    var newItemY = boxes.length ? boxes[boxes.length - 1].y - Math.random() * 150.0 - (init ? 100 : 0) : dimensions.height - 100 - (init ? 100 : 0);
-
-    var newItemX = Math.random() * (dimensions.width - 30);
-    var newItemW = Math.random() * 100 + 30;
-
-    if (boxes.length) {
-        // ensure the platform below this is not totally covered by this platform
-        var bottomLeftEdge = boxes[boxes.length - 1].x;
-        var bottomRightEdge = boxes[boxes.length - 1].x + boxes[boxes.length - 1].width;
-
-        var topLeftEdge = newItemX;
-        var topRightEdge = newItemX + newItemW;
-
-        if (topLeftEdge < bottomLeftEdge && topRightEdge > bottomRightEdge) {
-            // adjust one of the platforms
-            if (topLeftEdge < 150) {
-                newItemX = 20;
-            } else {
-                newItemX = 200;
-            }
-        }
-    }
-
-    if (!init && newItemY > 0)
-        newItemY = 0;
-
-    var count = boxes.length ? boxes[boxes.length - 1].count + 1 : 1;
-
-    if (count !== 0 && count % 20 === 0)
-        boxTypeOptions.push(boxTypes.ROCK);
-
-    var type;
-
-    do {
-        type = boxTypeOptions[Math.floor(Math.random() * boxTypeOptions.length)];
-    } while (boxes.length && type === boxTypes.ARROW && boxes[boxes.length - 1].type === boxTypes.ARROW);
-
-    var height = 5;
-
-    var isTopSolid,
-        isBottomSolid,
-        isLeftSolid,
-        isRightSolid,
-        adjustPlayer;
-
-    switch (type) {
-        case boxTypes.ROCK:
-            isTopSolid = true;
-            isBottomSolid = true;
-            isLeftSolid = true;
-            isRightSolid = true;
-            adjustPlayer = true;
-            break;
-        case boxTypes.ARROW:
-            isTopSolid = true;
-            isBottomSolid = true;
-            isLeftSolid = true;
-            isRightSolid = true;
-            adjustPlayer = false;
-            // arrows have specific requirements
-            newItemW = 50;
-            height = 38;
-
-            if (newItemX < 0)
-                newItemX = 0;
-            else if (newItemX >= dimensions.width - 50)
-                newItemX = dimensions.width - 50;
-            break;
-        default:
-            isTopSolid = false;
-            isBottomSolid = false;
-            isLeftSolid = false;
-            isRightSolid = false;
-            adjustPlayer = false;
-            break;
-    }
-
-    boxes.push({
-        x: newItemX,
-        y: newItemY,
-        width: newItemW,
-        height: height,
-        count: count,
-        type: type,
-        isTopSolid: isTopSolid,
-        isBottomSolid: isBottomSolid,
-        isLeftSolid: isLeftSolid,
-        isRightSolid: isRightSolid,
-        adjustPlayer: adjustPlayer
-    });
-}
-
-// seed the screen
-for (var i = 0; i < 3; i++)
-    generatePlatform(true);
-
-module.exports = {
-    boxTypes: boxTypes,
-    boxTypeOptions: boxTypeOptions,
-    generatePlatform: generatePlatform,
-    boxes: boxes
-};
-},{"./dimensions":2}],11:[function(require,module,exports){
-// normal squid
-var squidNeutral = new Image();
-
-squidNeutral.src = "/css/img/squid.png";
-squidNeutral.width = 30;
-squidNeutral.height = 35;
-
-// focus 1 squid
-var squidFocus1 = new Image();
-
-squidFocus1.src = "/css/img/squidJump1.png";
-squidFocus1.width = 30;
-squidFocus1.height = 35;
-
-// focus 2 squid
-var squidFocus2 = new Image();
-
-squidFocus2.src = "/css/img/squidJump2.png";
-squidFocus2.width = 30;
-squidFocus2.height = 35;
-
-// focus 3 squid
-var squidFocus3 = new Image();
-
-squidFocus3.src = "/css/img/squidJump3.png";
-squidFocus3.width = 30;
-squidFocus3.height = 35;
-
-// dead squid
-var squidDead = new Image();
-
-squidDead.src = "/css/img/squidDead.png";
-squidDead.width = 30;
-squidDead.height = 35;
-
-// arrow
-var arrow = new Image();
-
-arrow.src = "/css/img/arrow.png";
-arrow.width = 350;
-arrow.height = 38;
-
-
-// jump splash
-var splashVariant = Math.floor(Math.random() * 7);
-
-var splash = new Image();
-splash.src = "/css/img/splash" + splashVariant + ".png";
-splash.width = 1560;
-splash.height = 57;
-splash.className = "splashImage";
-
-module.exports = {
-    squidNeutral: squidNeutral,
-    squidFocus1: squidFocus1,
-    squidFocus2: squidFocus2,
-    squidFocus3: squidFocus3,
-    squidDead: squidDead,
-    splash: splash,
-    arrow: arrow
-};
-},{}],12:[function(require,module,exports){
-const keys = [];
-
-module.exports = {
-    keys: keys
-};
-},{}]},{},[7]);
