+++ src/userInput.js
@@ -1,0 +1,5 @@
+const keys = [];
+
+module.exports = {
+    keys: keys
+};
